% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_claim.R
\name{check_claim}
\alias{check_claim}
\title{Create claim boolean matrix}
\usage{
check_claim(code, target)
}
\arguments{
\item{code}{is a kcd code regular expression vector}

\item{target}{is a kcd code vector}
}
\description{
you can create the first claim matrix considering insured id.
}
\keyword{boolean}
\keyword{claim}
\keyword{matrix}
