% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_claim_1st.R
\name{check_claim_1st}
\alias{check_claim_1st}
\title{Create first claim boolean matrix by insured id from claim boolean matrix}
\usage{
check_claim_1st(data, id)
}
\arguments{
\item{data}{is a boolean matrix (claim kcd code exists or not)}

\item{id}{is a insured id numeric vector}
}
\description{
you can create the 1st claim matrix considering insured id.
}
\keyword{1st}
\keyword{claim}
