% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_claim_kcd.R
\name{check_claim_kcd}
\alias{check_claim_kcd}
\title{Create each rider's claim kcd code matrix}
\usage{
check_claim_kcd(code, target, cores)
}
\arguments{
\item{code}{is a kcd code regular expression vector}

\item{target}{is a claim kcd code vector}

\item{cores}{is a number of cores for parallel processing}
}
\description{
you can create each rider's claim kcd code matrix (from the kcd code matrix).
}
\keyword{claim}
\keyword{code}
\keyword{kcd}
