% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_claim_ratio.R
\name{check_claim_ratio}
\alias{check_claim_ratio}
\title{Create claim amount ratio matrix by claim year}
\usage{
check_claim_ratio(data, ratio)
}
\arguments{
\item{data}{is a claim year matrix}

\item{ratio}{is a claim amount ratio vector}
}
\description{
you can create the claim amount ratio matrix from the claim year matrix.
}
\keyword{amount}
\keyword{claim}
\keyword{ratio}
