% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_claim_rider.R
\name{check_claim_rider}
\alias{check_claim_rider}
\title{Create claim rider matrix}
\usage{
check_claim_rider(rider, code, target)
}
\arguments{
\item{rider}{is a rider vector}

\item{code}{is a kcd code regular expression vector}

\item{target}{is a claim kcd code vector}
}
\description{
you can create the claim rider matrix (from the claim boolean matrix included in the function).
}
\keyword{amount}
\keyword{claim}
\keyword{rider}
