% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_claim_year.R
\name{check_claim_year}
\alias{check_claim_year}
\title{Create year matrix from claim date vector}
\usage{
check_claim_year(date, origin, m)
}
\arguments{
\item{date}{is a claim date vector}

\item{origin}{is an orgin date of virtual underwriting}

\item{m}{is a length of riders}
}
\description{
you can create claim year matrix from virtual underwritng date
}
\keyword{claim}
\keyword{year}
