% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_first_claim.R
\name{check_first_claim}
\alias{check_first_claim}
\title{Create first claim boolean matrix by insured id from claim boolean matrix}
\usage{
check_first_claim(id, data)
}
\arguments{
\item{id}{is a insured id numeric vector}

\item{data}{is a boolean matrix (claim kcd code exists or not)}
}
\description{
you can create the first claim matrix considering insured id.
}
\keyword{claim}
\keyword{first}
