% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chk_clm.R
\name{chk_clm}
\alias{chk_clm}
\title{Create claim boolean matrix}
\usage{
chk_clm(code, target)
}
\arguments{
\item{code}{is a kcd code regular expression vector}

\item{target}{is a kcd code vector}
}
\description{
you can create the first claim matrix considering insured id.
}
\keyword{boolean}
\keyword{claim}
\keyword{matrix}
