% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chk_clm_1st.R
\name{chk_clm_1st}
\alias{chk_clm_1st}
\title{Create 1st claim boolean matrix by insured id from claim boolean matrix}
\usage{
chk_clm_1st(data, id)
}
\arguments{
\item{data}{is a boolean matrix (claim kcd code exists or not)}

\item{id}{is a insured id numeric vector}
}
\description{
you can create the 1st claim matrix considering insured id.
}
\keyword{1st}
\keyword{claim}
