% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chk_clm_mon.R
\name{chk_clm_mon}
\alias{chk_clm_mon}
\title{Create month matrix from claim date vector}
\usage{
chk_clm_mon(cdate, origin, m)
}
\arguments{
\item{origin}{is an orgin date of virtual underwriting}

\item{m}{is a length of riders}

\item{date}{is a claim date vector}
}
\description{
you can create claim month matrix from virtual underwritng date
}
\keyword{claim}
\keyword{year}
