% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chk_clm_rat.R
\name{chk_clm_rat}
\alias{chk_clm_rat}
\title{Create claim amount ratio matrix by claim year}
\usage{
chk_clm_rat(data, ratio)
}
\arguments{
\item{data}{is a claim year matrix}

\item{ratio}{is a claim amount ratio vector}
}
\description{
you can create the claim amount ratio matrix from the claim year matrix.
}
\keyword{amount}
\keyword{claim}
\keyword{ratio}
