% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chk_clm_rdr.R
\name{chk_clm_rdr}
\alias{chk_clm_rdr}
\title{Create claim rider vector}
\usage{
chk_clm_rdr(rider, code, target)
}
\arguments{
\item{rider}{is a rider vector}

\item{code}{is a kcd code regular expression vector}

\item{target}{is a claim kcd code vector}
}
\description{
you can create the claim rider vector (from the claim boolean matrix included in the function).
}
\keyword{amount}
\keyword{claim}
\keyword{rider}
