% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chk_prm.R
\name{chk_prm}
\alias{chk_prm}
\title{Create premium matrix by gender, age, grade}
\usage{
chk_prm(rsk_info, clm_info, old, gnd, grd, yrs)
}
\arguments{
\item{rsk_info}{is a risk information file}

\item{clm_info}{is a claim information file}

\item{old}{is an age}

\item{gnd}{is an gender}

\item{grd}{is an grade}

\item{yrs}{is a duration of observation}
}
\description{
you can create premium matrix by gender, age, grade
}
\keyword{matrix}
\keyword{premium}
