% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chk_prm_cnt.R
\name{chk_prm_cnt}
\alias{chk_prm_cnt}
\title{Create number of premium payment matrix by id}
\usage{
chk_prm_cnt(info, data, origin, yrs)
}
\arguments{
\item{info}{is a claim information file}

\item{data}{is a claim history file}

\item{origin}{is an origin date}

\item{yrs}{is a payment duration}
}
\description{
you can create number of premium payment matrix
}
\keyword{matrix}
\keyword{number}
\keyword{of}
\keyword{payment}
\keyword{premium}
