% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clm_sim.R
\name{clm_sim}
\alias{clm_sim}
\title{Claim simulation}
\usage{
clm_sim(info, data, origin)
}
\arguments{
\item{info}{is a claim information file}

\item{data}{is a claim history file}

\item{origin}{is an origin date}
}
\description{
Claim simulation by specific files (info, data)
}
\keyword{claim}
\keyword{simulation}
