% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clm_sim_hos.R
\name{clm_sim_hos}
\alias{clm_sim_hos}
\title{Create hospitalization days simulation matrix}
\usage{
clm_sim_hos(data, target)
}
\arguments{
\item{data}{is a hospitalization simulation matrix (through the clm_sim function)}

\item{target}{is a hospitalization days vector}
}
\description{
you can create the hospitalization days simulation matrix from hospitalization simulation matrix
}
\keyword{days}
\keyword{hospitalization}
\keyword{simulation}
