% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmb_clm_1st.R
\name{cmb_clm_1st}
\alias{cmb_clm_1st}
\title{Combine claim matrix with 1st claim matrix by one time payment vector}
\usage{
cmb_clm_1st(clm, clm_1st, otime)
}
\arguments{
\item{clm}{is a claim matrix}

\item{clm_1st}{is a 1st claim matrix}

\item{otime}{is a one time benefit vector (one-time benefits or continuous benefits)}
}
\description{
you can combine claim matrix with 1st claim matrix
}
\keyword{1st}
\keyword{and}
\keyword{claim}
\keyword{combine}
