% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cst_unq_kcd.R
\name{cst_unq_kcd}
\alias{cst_unq_kcd}
\title{Reshape kcd code column long-to-wide version by each var_id}
\usage{
cst_unq_kcd(data, var_id, target, prefix = "var", glue = TRUE)
}
\arguments{
\item{data}{is a claim data}

\item{var_id}{is var_id}

\item{target}{is a kcd column variable}

\item{prefix}{is a prefix for new kcd columns}

\item{glue}{is to glue all new kcd columns}
}
\description{
you can transfrom kcd code columns long-to-wide version by each var_id
}
\keyword{code}
\keyword{column}
\keyword{kcd}
\keyword{long-to-wide}
\keyword{reshape}
