% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cut_age.R
\name{cut_age}
\alias{cut_age}
\title{Categorize the age variable}
\usage{
cut_age(data, var, interval, right = FALSE)
}
\arguments{
\item{data}{is a claim data}

\item{var}{is a age column variables}

\item{interval}{is an interval number}

\item{right}{is a limit direction (default FALSE)}
}
\description{
you can categorize the age variable.
}
\keyword{hospitalization}
