% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prm_sim.R
\name{prm_sim}
\alias{prm_sim}
\title{Create premium payment simulation data}
\usage{
prm_sim(rsk_info, clm_info, data, origin, yrs)
}
\arguments{
\item{rsk_info}{is a risk information file}

\item{clm_info}{is a claim information file}

\item{data}{is a claim history file}

\item{origin}{is an origin date}

\item{yrs}{is a payment duration}
}
\description{
you can create number of premium payment matrix
}
\keyword{payment}
\keyword{premium}
\keyword{simulation}
