% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spl_hos_day.R
\name{spl_hos_day}
\alias{spl_hos_day}
\title{Split hospitalization days by origin date (virtual uw date)}
\usage{
spl_hos_day(data, var_from, var_to, origin, all = TRUE)
}
\arguments{
\item{data}{or data.table with hospitalization data}

\item{var_from}{start date column variable}

\item{var_to}{end date column variable}
}
\description{
you can create new hospitalization days data frame divided by origin date.
}
\keyword{by}
\keyword{date}
\keyword{days}
\keyword{hospitalization}
\keyword{origin}
\keyword{split}
