% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spl_unq_kcd.R
\name{spl_unq_kcd}
\alias{spl_unq_kcd}
\title{Split each kcd codes and increase the data by the number of each kcd code}
\usage{
spl_unq_kcd(data, var)
}
\arguments{
\item{data}{is a claim data}

\item{var}{is a kcd column variable}
}
\description{
you can create increased data splitted by each unique kcd code
}
\keyword{codes}
\keyword{each}
\keyword{kcd}
\keyword{split}
