% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srch_kcd.R
\name{srch_kcd}
\alias{srch_kcd}
\title{Search description of kcd}
\usage{
srch_kcd(kcd, lang = "kr")
}
\arguments{
\item{string}{A kcd code or regular expression}
}
\description{
you can search kcd and find its description in the kcd book.
}
\examples{
srch_kcd('J00')
srch_kcd('^M51$')
}
