% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unq_hos_day.R
\name{unq_hos_day}
\alias{unq_hos_day}
\title{Calculate unique hospitalization days}
\usage{
unq_hos_day(data, var_id, var_from, var_to)
}
\arguments{
\item{data}{or data.table with hospitalization data like id, icd code, start date, end date.}

\item{var_id}{grouping column variables}

\item{var_from}{start date column variable}

\item{var_to}{end date column variable}
}
\description{
you can calculate the hospitalization days by several groups like id, icd code
considering duplicated days.
}
\keyword{hospitalization}
