% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unq_sur_cnt.R
\name{unq_sur_cnt}
\alias{unq_sur_cnt}
\title{Calculate unique surgery counts}
\usage{
unq_sur_cnt(data, var_id, var_to)
}
\arguments{
\item{data}{is claim data with id, icd code, start date, end date.}

\item{var_id}{is the grouping column variables}

\item{var_to}{is the surgery date column variable}
}
\description{
you can calculate the number of surgery by several groups like id, icd code
considering duplicated days.
}
\keyword{counts}
\keyword{surgery}
