% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uniq_hos_day.R
\name{uniq_hos_day}
\alias{uniq_hos_day}
\title{uniq_hos_day Function}
\usage{
uniq_hos_day(data, var_id, var_from, var_to)
}
\arguments{
\item{data.frame}{or data.table with hospitalization data like id, icd code, start date, end date.}
}
\description{
you can calculate the hospitalization days by several groups like id, icd code
considering duplicated days.
}
\keyword{hospitalization}
